infrastructure/commons/datadog/requirements.txt
datadog-api-client~=1.5.0
PyYAML~=6.0
